TT#35550 Update GRML binary files to GRML version 2018-04-11 (autobuild-build2338)
The previous version 2017-05-31 is not stable enough to be used in production,
even SSH to the machine which is being installed doesn't work well. One more
example here is lack of fix for already reported interactive scripts regression:
https://github.com/grml/grml-autoconfig/commit/55ce0afb19f05761a5a24ce6e20a5ce6a7aa58dd

The rough test for the latest 'testing' version from today shows much better stability.
Using it and "the official GRML release will follow if we find no issues" (c) mprokop@.

Also the full version size has been increased on 30% since GRML 2014-11-14:

> -r--r--r-- 1 sipwise sipwise 436916224 May 31  2017 ./grml20141114/live/grml64-full/grml64-full.squashfs
> -r--r--r-- 1 sipwise sipwise 571748352 May 31  2017 ./grml20170531/live/grml64-full/grml64-full.squashfs
> -r--r--r-- 1 sipwise sipwise 605184000 Apr 11 06:12 ./grml20180411/live/grml64-full/grml64-full.squashfs

So we are trying to migrate here to the small version:

> -r--r--r-- 1 sipwise sipwise 268627968 Apr 11 06:37 grml20180411_small/live/grml64-small/grml64-small.squashfs

Initial tests shows small modifications required for deployemnt.sh.

Change-Id: I2231aaecd20fb1f6847e19c018ec20d754dfa2fa